+++ Resources/Prototypes/_Nuclear14/Entities/Structures/Walls/rock.yml
@@ -11,7 +11,6 @@
   - type: IconSmooth
     key: walls
     mode: NoSprite
-  - type: SmoothEdge
   - type: Icon
     sprite: Structures/Walls/rock.rsi
     state: rock_asteroid
@@ -188,7 +187,6 @@
   - type: IconSmooth
     key: walls
     mode: NoSprite
-  - type: SmoothEdge
   - type: Sprite
     sprite: Structures/Walls/rock.rsi
     layers:
@@ -267,7 +265,6 @@
     - type: Icon
       sprite: Structures/Walls/rock.rsi
       state: rock
-    - type: SmoothEdge
     - type: Sprite
       sprite: Structures/Walls/rock.rsi
       layers:
